Modify sources.py script for output formatting
The script should not overwrite data/news_sources/sources_list.md since
there are notes for choosing which sources. The script will also output
a list of the ID of the sources in json format, so that it can be easily
used later to get articles with the sources we want.